separate tasks into two different tasks + add yield_per
