+++ tasks/__init__.py
@@ -1,7 +1,12 @@
 from app import celery_app
 from tasks.ai_pr_review import ai_pr_view_task
 from tasks.backfill_commit_data_to_storage import backfill_commit_data_to_storage_task
-from tasks.backfill_gh_app_installations import backfill_gh_app_installations_name
+from tasks.backfill_existing_gh_app_installations import (
+    backfill_existing_gh_app_installations_name,
+)
+from tasks.backfill_owners_without_gh_app_installations import (
+    backfill_owners_without_gh_app_installations_name,
+)
 from tasks.brolly_stats_rollup import brolly_stats_rollup_task
 from tasks.bundle_analysis_notify import bundle_analysis_notify_task
 from tasks.bundle_analysis_processor import bundle_analysis_processor_task
